Refactor tests to reduce duplication
The stubs to create the app and versionfile in tests/test_app.py
would also be needed in every other test file that wanted to test
the Flask app. This commit extracts those utility functions into
their own file which can be imported into any test that needs them.
This commit also renames 'test_app.py' to 'tests_dockerflow.py' to
better represent what that file is testing.
