+++ dev-requirements.txt
@@ -25,3 +25,6 @@ mccabe==0.6.1 \
 py==1.4.33 \
     --hash=sha256:81b5e37db3cc1052de438375605fb5d3b3e97f950f415f9143f04697c684d7eb \
     --hash=sha256:1f9a981438f2acc20470b301a07a496375641f902320f70e31916fe3377385a9
+requests-mock==1.3.0 \
+    --hash=sha256:23edd6f7926aa13b88bf79cb467632ba2dd5a253034e9f41563f60ed305620c7
+
