+++ tests/utils.py
@@ -1,0 +1,33 @@
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this
+# file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
+import json
+
+import pytest
+
+from landoapi.app import create_app
+
+
+@pytest.fixture
+def versionfile(tmpdir):
+    """Provide a temporary version.json on disk."""
+    v = tmpdir.mkdir('app').join('version.json')
+    v.write(
+        json.dumps(
+            {
+                'source': 'https://github.com/mozilla-conduit/lando-api',
+                'version': '0.0.0',
+                'commit': '',
+                'build': 'test',
+            }
+        )
+    )
+    return v
+
+
+@pytest.fixture
+def app(versionfile):
+    """Needed for pytest-flask."""
+    app = create_app(versionfile.strpath)
+    return app.app
